code cleanup

+++ src/citation/related_models/citation_project_model.py
@@ -4,7 +4,6 @@ from django.db import models
 from researchhub_access_group.constants import ADMIN, EDITOR, VIEWER
 from researchhub_access_group.related_models.permission_model import Permission
 from user.models import Organization
-from user.related_models.user_model import User
 from utils.models import DefaultAuthenticatedModel
 
 
@@ -44,9 +43,7 @@ class CitationProject(DefaultAuthenticatedModel):
         for editor_id in editor_ids:
             editor_exists = self.permissions.has_editor_user(editor_id)
             if not editor_exists:
-                self.permissions.create(
-                    access_type=EDITOR, user=User.objects.get(id=editor_id)
-                )
+                self.permissions.create(access_type=EDITOR, user=editor_id)
         return True
 
     def add_viewers(self, viwer_ids=[]):
@@ -74,9 +71,7 @@ class CitationProject(DefaultAuthenticatedModel):
 
     def remove_editors(self, editor_ids):
         for editor_id in editor_ids:
-            self.permissions.filter(
-                access_type=EDITOR, user=User.objects.get(id=editor_id)
-            ).all().delete()
+            self.permissions.filter(access_type=EDITOR, user=editor_id).all().delete()
         return True
 
     def remove_viewers(self, viewer_ids):
